requirements.txt
numpy~=1.23.1
PyYAML~=6.0
torch~=1.12.0
Shapely~=1.8.2